Minor change in getConfig
Some people don't or can't load a ```Config/cloud_files.php``` configuration file. So there have to be a way to avoid loading (inside the plugin).

For example, if I have a configuration file in another place that is not app/Config and I load that config, the plugin should not try to load it again (and possibly merge it).